Remove the FD solver, add the potential equation
Current changes incorporate the change to remove the FD solver for
radial integration. This is necessary since the FD solver is not
necessary for solving TEHD flow and works well with Chebyshev
polynomials.

The electric potential is implemented via new file updateV. The
according change to solve the electric field are also done, especially
for solving the temperature gradient and electric potential v in the DEP
force.